Move create_score_element() to typing/element.py
